+++ src/components/auto-mode-button/modes/error/Dialog.vue
