+++ src/components/auto-mode-button/modes/off/Dialog.vue
