+++ src/components/auto-mode-button/modes/on/Dialog.vue
